requirements.txt
python-dotenv
PyMuPDF
spacy
arxiv
colorama
deepl
googletrans==4.0.0-rc1